server.py setuptools entry point
